user story 2 complete

+++ JavaWebServiceGroupB/src/main/java/com/kainos/ea/employee_stuff/DeliveryEmployee.java
@@ -5,7 +5,7 @@ public class DeliveryEmployee extends Employee{
     private String image;
     private String favTech;
 
-    public DeliveryEmployee(short number, int salary, String firstName, String lastName, String bankAccountNumber, String niNumber, String phoneNumber, String email, String cv, String image, String favTech) {
+    public DeliveryEmployee(short number, double salary, String firstName, String lastName, String bankAccountNumber, String niNumber, String phoneNumber, String email, String cv, String image, String favTech) {
         super(number, salary, firstName, lastName, bankAccountNumber, niNumber, phoneNumber, email);
         this.cv = cv;
         this.image = image;
